requirements.txt
python-dotenv>=1.0.0
streamlit>=1.30.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langgraph>=0.2.0
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
python-dateutil>=2.8.0
selenium>=4.15.0
PyMuPDF>=1.23.0
